implement cookie-based authentication and fix auth handling on frontend

+++ src/lib/actions/connection.actions.js
@@ -1,12 +1,24 @@
 "use server"
 
 import axios from 'axios';
+import { cookies } from 'next/headers'
 
 const BASE_URL = process.env.BASE_URL
 
 export async function getLinkEmailToPartnerAction() {
     try {
-        const { data } = await axios.get(`${BASE_URL}/api/v1/aerozone/connection/all`);
+        const cookieStore = await cookies();
+        const token = cookieStore.get('access_token')?.value;
+
+        if (!token) {
+            throw new Error('No token – the user is not logged in.');
+        }
+        const { data } = await axios.get(`${BASE_URL}/api/v1/aerozone/connection/all`, {
+            headers: {
+                'Cookie': `access_token=${token}`
+            },
+            withCredentials: true,
+        });
         return data;
     } catch (error) {
         console.error('Kapcsolatok lekérdezési hiba:', error);
@@ -32,7 +44,20 @@ export async function createLinkEmailToPartnerAction(prevState, formData) {
     }
 
     try {
-        const res = await axios.post(`${BASE_URL}/api/v1/aerozone/connection/create?partner_id=${partner_id}&email_id=${email_id}`);
+        const cookieStore = await cookies();
+        const token = cookieStore.get('access_token')?.value;
+
+        if (!token) {
+            throw new Error('No token – the user is not logged in.');
+        }
+        console.log(partner_id, email_id);
+        const res = await axios.post(`${BASE_URL}/api/v1/aerozone/connection/create?partner_id=${partner_id}&email_id=${email_id}`, {}, {
+            headers: {
+                'Cookie': `access_token=${token}`
+            },
+            withCredentials: true,
+        });
+
         return {
             success: true,
             message: 'A kapcsolat sikeresen létrejött!',
@@ -52,11 +77,21 @@ export async function createLinkEmailToPartnerAction(prevState, formData) {
 
 export async function deleteConnectionAction(emailId, partnerId) {
     try {
+        const cookieStore = await cookies();
+        const token = cookieStore.get('access_token')?.value;
+
+        if (!token) {
+            throw new Error('No token – the user is not logged in.');
+        }
         const { data } = await axios.delete(`${BASE_URL}/api/v1/aerozone/connection/delete`, {
             params: {
                 email_id: emailId,
                 partner_id: partnerId
-            }
+            },
+            headers: {
+                'Cookie': `access_token=${token}`
+            },
+            withCredentials: true,
         });
         return data;
     } catch (error) {
